Start work on alternative client
This client will have nicer user interface. The client allows loading
plugins from predefined directories so that implementing internal
extensions is possible.

So far, listing, retrieving, creating and updating releases is possible.

Ultimately, the PDCClient class should be moved to a separate module, so
that it is accessible for other users. It abstracts the reading the
client configuration file and obtaining token.

Another thing that must be improved is the help - currently all commands
with help are shown with --help, and commands without help are not shown
at all. This should be split into --help and --help-admin.