Add commands for working with rpms
This patch adds commands for listing, showing, creating and updating
RPMs. When showing info about particular RPM or updating, the database
identifier is used. This could be replaced with NEVRA, but would require
another dependency on something that can parse it.

Unrelated to RPMs, this patch also modifies the way errors are presented
to the user. This does not handle all possible failures.

There are now also two methods that might simplify processing arguments
from command line in that it is sufficient to declare them only once.
The original manual way is still available to handle more complicated
cases.

JIRA: PDC-1023