Make changesets capable of handling dates.
I ran into an issue where a new model in a new app was using a `date`
field.  I could create and retrieve these resources, no problem.  But,
when I went to PATCH them, it would fail.  The error turned up to be in
this stretch of code, where JSON-encoded versions of the objects are
compared to track the audit trail.. and the default json encoder doesn't
know how to serialize date objects by default.

This adds a date-aware JSON serializer that gets things working again.