Fix displaying deleted composes
When a compose is marked as deleted, it should not show up by default
in some places. The compose list has a filter, so there it's ok.

This patch fixes and tests that compose_set in /releases/ API does not
include deleted composes (either directly connected or linked).

Also the lower level method used to get latest compose for a release is
tested to not include deleted ones. This should fix RPM Mapping API
end-point.

The APIs to get compose by release or product version and RPM NVR are
updated to not return deleted composes.

JIRA: PDC-2533